refact(code validator): change code structure and rename components
